DRILL-7744: Move Filters from HTTP Storage Plugin to Drill Core
