+++ exec/java-exec/src/main/java/org/apache/drill/exec/store/base/filter/ExprNode.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.drill.exec.store.http.filter;
+package org.apache.drill.exec.store.base.filter;
 
 import java.util.List;
